Conditionally display template details link (#1957)
If the policy framework can not create the policy template (because the
CRD does not exist, or if the template is invalid), then the link that
was previously added would go to a page with a not found error. This
change detects most situations where that would occur, and skips adding
the link.

Refs:
 - https://github.com/stolostron/backlog/issues/24746

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>