Fix the related objects view YAML link in hosted mode
In hosted mode, the cluster where the policy templates exist is
different than the managed cluster that the policy templates apply to.
This commit accounts for this when the Policy addon is installed in
hosted mode.

Now the "View yaml" link will search for the related object on the
managed cluster instead of the hosting cluster.

Relates:
https://github.com/stolostron/backlog/issues/25103
https://bugzilla.redhat.com/show_bug.cgi?id=2118338

Signed-off-by: mprahl <mprahl@users.noreply.github.com>