Add alert to Host Inventory card when no OpenShift images are available (#2104)
When installing a cluster using host inventory, if there are no clusterImageSets
available, user is not able to select and OpenShift version and gets blocked in
the cluster creation wizard. This change disables the Host Inventory
card and adds hint on how to add release images.

Fixes BZ2124581
Cherry-picked from 0d3a2365

Signed-off-by: Jiri Tomasek <jtomasek@redhat.com>

Signed-off-by: Jiri Tomasek <jtomasek@redhat.com>