Submariner: hide infra-related fields for ARO/ROSA
For ARO and ROSA platforms Submariner should not configure
anything in the infrastructure, so there is no need to
display the infrastructure related fields.

This PR updates the user interface to hide the
infra-related fields for the case
of managed platforms and also set loadBalancerEnable
default to true for managed platforms.

Signed-off-by: Yossi Boaron <yboaron@redhat.com>