Protect fewer fields in create policy wizard
Previously, when `metadata.uid` was present in an object in the editor,
that field and its siblings would always be readonly. This prevented
changes to fields that should not be changed when editing a resource,
but also made it impossible to paste an existing resource into a create
wizard and edit those fields.

Now, this behavior is configurable in the SyncEditorProps. By default,
the old behavior will be used. The Create Policy wizard opts in and uses
the new behavior.

Refs:
 - https://issues.redhat.com/browse/ACM-2188

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>