MGMT-13551 Move installAttemptsLimit control to automationControlData
`installAttemtpsLimit` control flag is currently only used by `onChangeAutomationTemplate`
function inside automationControlData. The function relies on the control to be defined
in the controlData otherwise an error occurs when trying to set installAttemtsLimit.active
value. Since the installAttemptsLimit control was defined in clusterDetailsControlData, for the
wizards (e.g. CIM or AI) which don't use this step the error occured.

To fix this, we're moving the installAttemptsLimit control to automationControlData which is only
controlData which actually uses that flag.

Cherry-picked from 405b2654ef1b1f0ad19dafda8a177d72690961c5

Signed-off-by: Jiri Tomasek <jtomasek@redhat.com>