Fix the infinite loop of calling fireManagedClusterView
The PolicyTemplateDetails component was responsible for using
ManagedClusterView to retrieve the policy template YAML. Recently, it
started setting the policy template YAML on the parent component of
PolicyTemplateDetailsPage. It did this with the props.setParentTemplate
function, but a dependency on props in useEffect was causing the
function to be executed constantly.

There are workarounds, but the cleanest solution is to have
PolicyTemplateDetailsPage fetch the policy template YAML and pass the
YAML down as a prop.

Relates:
https://issues.redhat.com/browse/ACM-12016

Signed-off-by: mprahl <mprahl@users.noreply.github.com>