Reduce renders and race conditions in policy template details
This refactors policy template details components to reduce renders.

Additionally, fireManagedClusterView was improved to be able to
explicitly name the created ManagedClusterView objects to prevent
conflicts and properly ignore duplicate fireManagedClusterView calls.
This mostly helps local development or when the console is run in strict
mode and hooks are run twice. It also helps when two separate browser
windows are looking at the same policy template, especially Kyverno ones
with many audit violations.

Lastly, some loading states were being set to false after the
fireManagedClusterView promise was created but before it resolved. This
is now fixed to wait for the promise to be resolved.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>