Initial discovered policy resources page
The page is largely based on `PolicyTemplateDetails.tsx`. The work of
sorting through the search response is delegated to the worker thread.

Because not all the information we want to display on the page is
suitable to be saved in Search (for example, we want the 'Reason'
messages and `Diff` information when it is present on a
ConfigurationPolicy), the page uses ManagedClusterViews for each
displayed policy to get some more information. Those are only fired when
a related resource for that policy should be displayed, and they are
saved to prevent duplicated requests.

Refs:
 - https://issues.redhat.com/browse/ACM-20098

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>