Update compliance operator install policy (#5284)
This pulls in the change from
https://github.com/open-cluster-management-io/policy-collection/pull/533

From that PR: "Previously, the template used fromClusterClaim, but if
the cluster claim could not be found, the policy would become
noncompliant. Now it uses lookup which will return an empty string if
the resource can't be found. This allows the template to work in more
cases."

Refs:
 - https://issues.redhat.com/browse/ACM-25409

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
Co-authored-by: Justin Kulikauskas <jkulikau@redhat.com>